pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.1"]
build-backend = "poetry.core.masonry.api"

[project]
name = "keystone-api"
version = "0.0.0"
readme = "README.md"
description = "A REST API for managing user resource allocations on HPC systems."
authors = [{ name = "Better HPC LLC" }]
keywords = ["Better", "HPC", "keystone", "api"]
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Topic :: Internet",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Typing :: Typed"
]
requires-python = ">=3.11,<4"
dependencies = [
    "coverage==7.9.1",
    "django==5.2.4",
    "django-auditlog==3.2.0",
    "django-celery-beat==2.8.1",
    "django-celery-results==2.6.0",
    "django-cors-headers==4.7.0",
    "django-environ==0.12.0",
    "django-filter==25.1",
    "django-health-check==3.20.0",
    "django-jazzmin==3.0.1",
    "django-prometheus==2.4.1",
    "djangorestframework==3.16.0",
    "drf_spectacular[sidecar]==0.28.0",
    "factory-boy==3.3.3",
    "Faker==37.4.0",
    "flower==2.0.1",
    "jinja2==3.1.6",
    "pillow==11.3.0",
    "psycopg2-binary==2.9.10",
    "pyyaml==6.0.2",
    "redis==6.2.0",
    "servestatic==3.1.0",
    "tqdm==4.67.1",
    "uritemplate==4.2.0",
    "uvicorn==0.35.0",
]

[project.optional-dependencies]
ldap = ["django-auth-ldap"]
smtp = ["aiosmtpd"]
tests = ["coverage"]
all = ["django-auth-ldap", "aiosmtpd", "coverage"]

[tool.poetry.group.tests.dependencies]
coverage = { version = "*", optional = true }

[tool.poetry.group.ldap.dependencies]
django-auth-ldap = { version = "5.2.0", optional = true }

[tool.poetry.group.smtp.dependencies]
aiosmtpd = { version = "1.4.6", optional = true }

[tool.coverage.run]
branch = true
omit = [
    "*/tests/*",
    "*/migrations/*",
    "keystone_api/main/*",
]

[project.scripts]
keystone-api = "keystone_api.manage:main"
